[IMP] account: Purchase UOM on Vendor Bills
Before this commit, selecting a product on a vendor bill line would set the product sale uom as the line uom by default.
This behavior was changed to set the product purchase uom as the line uom by default.

The default uom behavior on Vendor Bills lines should be aligned with the default uom behavior on Purchase orders.

task-4012191

closes odoo/odoo#175235

X-original-commit: bff969d71360ac69bbfad87809ff15d96e5a6f9c
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>