[FIX] base: stdnum library timeout
stdnum library incorrectly sets zeep Transport timeout, resulting in
some requests hanging for 15 minutes. With this monkeypatch the timeout
will be set correctly. The monkeypatch is a full copy of the original
code except for adding operational_timeout to the Transport.

The monkeypatch can be removed when
arthurdejong/python-stdnum/issues/444 is resolved and the version is
upgraded

Related zeep github issue: mvantellingen/python-zeep#140
This fix was already commited but reverted due to an unexpected
side-effect:
Original commit: 11062f6
Revert: de9df2ba5c15

The issue originated from a dependency issue in the zeep library in
Ubuntu 22.04. The python3-platformdirs dependency is missing in there,
yet it is used in that version of zeep. This missing dependency used to
be hidden by the chains of try ... except ImporError of stdnum
That means that it's another import from the chain that was used. So to
reduce the possibility of introducing a similar bug and catch such
errors, the full code was copied.
See odoo/odoo#173359 for a full explanation of why the full code is
copied even though some libraries are not in the requirements.

opw-3980718

closes odoo/odoo#174340

X-original-commit: 6eaf590a8f4178974545b91c3d0a73b7d5d90ec8
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Joris Makauskis (jmak) <jmak@odoo.com>