[FIX] mrp: avoid changing the consumed qty
Backport of [1]

The commit says:
> The issue does not appear in 17.0 because, thanks to commit 5bb0f96:
The `manual_consumption` field of the stock moves related to a
manufacturing order is set to be True as soon as the quantity is changed
because of these lines:
https://github.com/odoo/odoo/blob/9a11717c17b860ec2f1b2e228517c0d3945c474a/addons/mrp/static/src/widgets/mrp_consumed.js#L25-L27

But, if the user changes that quantity through the detailed
operations, nothing will write on `manual_consumption`. As a result,
the user then ticks the _Consumed_ box, sets a producing qty and...
The bug is back: the consumed qty of the components will be updated.

In the above use case, if [1] is applied, when the user ticks the
box, it will trigger the onchange and flag `manual_consumption`.

[1] 8aa7800fdf52b765b2b08b33aed1b963ff141401

OPW-4048258

closes odoo/odoo#174419

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Co-authored-by: lase-odoo <lase@odoo.com>